+++ macos/Classes/FFmpegKitFlutterPlugin.h
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+#import <FlutterMacOS/FlutterMacOS.h>
+
+@interface FFmpegKitFlutterPlugin : NSObject <FlutterPlugin>
+@end